Refactor app handlers
