+++ pushmanager/pushmanager_main.py
@@ -3,18 +3,22 @@ from __future__ import with_statement
 import os
 import tornado.httpserver
 import tornado.web
-import urlparse
 
 from pushmanager.core.application import Application
-from pushmanager.core.auth import authenticate
 import pushmanager.core.db as db
 from pushmanager.core.git import GitQueue
 from pushmanager.core.mail import MailQueue
 from pushmanager.core.rb import RBQueue
-from pushmanager.core.requesthandler import RequestHandler
 from pushmanager.core.settings import Settings
 from pushmanager.core.util import get_servlet_urlspec
 from pushmanager.core.xmppclient import XMPPQueue
+
+from pushmanager.handlers import CheckSitesBookmarkletHandler
+from pushmanager.handlers import CreateRequestBookmarkletHandler
+from pushmanager.handlers import LoginHandler
+from pushmanager.handlers import LogoutHandler
+from pushmanager.handlers import NullRequestHandler
+from pushmanager.handlers import RedirHandler
 
 from pushmanager.servlets.addrequest import AddRequestServlet
 from pushmanager.servlets.api import APIServlet
@@ -48,64 +52,6 @@ from pushmanager.servlets.verifyrequest import VerifyRequestServlet
 
 import pushmanager.ui_modules as ui_modules
 import pushmanager.ui_methods as ui_methods
-
-class NullRequestHandler(RequestHandler):
-    def get(self): pass
-    def post(self): pass
-
-class BookmarkletHandler(RequestHandler):
-    bookmarklet = None
-    def get(self):
-        if self.bookmarklet:
-            self.render(self.bookmarklet)
-
-class CreateRequestBookmarkletHandler(BookmarkletHandler):
-    bookmarklet = "create_request_bookmarklet.js"
-    # Create Request Bookmarklet is served from /bookmarklet to keep backwards compatibility.
-    url = r"/bookmarklet"
-
-class CheckSitesBookmarkletHandler(BookmarkletHandler):
-    bookmarklet = "check_sites_bookmarklet.js"
-    url = r"/checksitesbookmarklet"
-
-class LoginHandler(RequestHandler):
-    def get(self):
-        next_url = self.request.arguments.get('next', [None])[0]
-        if self.current_user:
-            return self.redirect(next_url or '/')
-        self.render("login.html", page_title="Login", errors=None, next_url=next_url)
-
-    def post(self):
-        next_url = self.request.arguments.get('next', [None])[0]
-        username = self.request.arguments.get('username', [None])[0]
-        password = self.request.arguments.get('password', [None])[0]
-
-        if self.current_user:
-            return self.redirect(next_url or '/')
-
-        if not username or not password:
-            return self.render("login.html", page_title="Login", next_url=next_url,
-                errors="Please enter both a username and a password.")
-        if not authenticate(username, password):
-            return self.render("login.html", page_title="Login", next_url=next_url,
-                errors="Invalid username or password specified.")
-
-        self.set_secure_cookie("user", username)
-        return self.redirect(next_url or "/")
-
-class LogoutHandler(RequestHandler):
-    def get(self):
-        self.clear_cookie("user")
-        return self.redirect("/")
-    post = get
-
-class RedirHandler(tornado.web.RequestHandler):
-    def get(self, path):
-        self.redirect(urlparse.urljoin(
-            'https://%s/' % Settings['main_app']['servername'],
-            path,
-        ), permanent=True)
-    post = get
 
 
 # Servlet dispatch rules
@@ -149,6 +95,7 @@ def get_url_specs():
                     MsgServlet):
         url_specs.append(get_servlet_urlspec(servlet))
     return url_specs
+
 
 class PushManagerApp(Application):
     name = "main"
