+++ pushmanager/tests/test_bookmarklet.py
@@ -1,8 +1,8 @@
 import contextlib
 import mock
 
-from pushmanager.pushmanager_main import CreateRequestBookmarkletHandler
-from pushmanager.pushmanager_main import CheckSitesBookmarkletHandler
+from pushmanager.handlers import CreateRequestBookmarkletHandler
+from pushmanager.handlers import CheckSitesBookmarkletHandler
 from pushmanager.testing.testservlet import AsyncTestCase
 import pushmanager.testing as T
